+++ lib/action_policy/rails/policy/instrumentation.rb
@@ -6,12 +6,19 @@ module ActionPolicy # :nodoc:
       # Add ActiveSupport::Notifications support.
       #
       # Fires `action_policy.apply_rule` event on every `#apply` call.
+      # Fires `action_policy.init` event on every policy initialization.
       module Instrumentation
-        EVENT_NAME = "action_policy.apply_rule"
+        INIT_EVENT_NAME = "action_policy.init"
+        APPLY_EVENT_NAME = "action_policy.apply_rule"
+
+        def initialize(*)
+          event = {policy: self.class.name}
+          ActiveSupport::Notifications.instrument(INIT_EVENT_NAME, event) { super }
+        end
 
         def apply(rule)
           event = {policy: self.class.name, rule: rule.to_s}
-          ActiveSupport::Notifications.instrument(EVENT_NAME, event) do
+          ActiveSupport::Notifications.instrument(APPLY_EVENT_NAME, event) do
             res = super
             event[:cached] = result.cached?
             event[:value] = result.value